refactor: move display manager to services and implement display loop functionality
